Take resources andd  add new  page with css and bootstratp
Get very cool  free html templets
delete old fav icon and css and replaced with the new.
Get the beginning of a html page and and in to index.html put there some css conection an bootrstrap an js files
make correction on a mean html in app.js putting self close img/ and hr add one more div and  fix all errors on a console.
Now the app is running excellent